requirements.txt
Django==1.7.1
Pygments==2.0.1
certifi==14.05.14
defusedxml==0.4.1
django-allauth==0.19.0
django-cors-headers==1.0.0
django-filter==0.9.2
django-rest-auth==0.3.4
djangorestframework==3.0.3
httpie==0.8.0
ipython==2.3.1
oauthlib==0.7.2
python3-openid==3.0.5
pyzmq==14.4.1
requests==2.5.0
requests-oauthlib==0.4.2
six==1.9.0
tornado==4.0.2
